Completed function that can find all useful drug combinations for given disease

+++ scripts/combineData.sh
@@ -3,6 +3,11 @@
 FILES=$(cat clean/splitData.txt)
 
 for file in $FILES; do
+    if [ -e $file ]; then
+        echo "File $file already exists"
+        exit 1
+    fi
+
     cat $file.zip.?? > "$file.zip"
     rm $file.zip.??
     gunzip -S .zip "$file.zip"
